移动 getPushProvider 到 PBHPushController
