+++ src/main/java/com/ghostchu/peerbanhelper/module/impl/webapi/PBHPushController.java
@@ -1,8 +1,12 @@
 package com.ghostchu.peerbanhelper.module.impl.webapi;
 
+import com.ghostchu.peerbanhelper.Main;
 import com.ghostchu.peerbanhelper.module.AbstractFeatureModule;
-import com.ghostchu.peerbanhelper.push.PushManager;
 import com.ghostchu.peerbanhelper.push.PushProvider;
+import com.ghostchu.peerbanhelper.push.impl.PushPlusPushProvider;
+import com.ghostchu.peerbanhelper.push.impl.ServerChanPushProvider;
+import com.ghostchu.peerbanhelper.push.impl.SmtpPushProvider;
+import com.ghostchu.peerbanhelper.push.impl.TelegramPushProvider;
 import com.ghostchu.peerbanhelper.text.Lang;
 import com.ghostchu.peerbanhelper.util.context.IgnoreScan;
 import com.ghostchu.peerbanhelper.web.JavalinWebContainer;
@@ -11,6 +15,7 @@ import com.ghostchu.peerbanhelper.web.wrapper.StdResp;
 import io.javalin.http.Context;
 import io.javalin.http.HttpStatus;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -21,8 +26,6 @@ import static com.ghostchu.peerbanhelper.text.TextManager.tl;
 public class PBHPushController extends AbstractFeatureModule {
     @Autowired
     private JavalinWebContainer webContainer;
-    @Autowired
-    private PushManager pushManager;
 
     @Override
     public boolean isConfigurable() {
@@ -47,7 +50,7 @@ public class PBHPushController extends AbstractFeatureModule {
 
     private void handlePushTest(Context ctx) {
         String name = ctx.pathParam("name");
-        PushProvider provider = pushManager.getPushProvider(name);
+        PushProvider provider = getPushProvider(name);
         if (provider == null) {
             ctx.status(HttpStatus.NOT_FOUND);
             return;
@@ -62,6 +65,33 @@ public class PBHPushController extends AbstractFeatureModule {
         } catch (Exception e) {
             ctx.json(new StdResp(false, tl(locale(ctx), Lang.PUSH_PROVIDER_TEST_FAILED, name, providerType, e), null));
         }
+    }
+
+    @Nullable
+    private PushProvider getPushProvider(String name) {
+        if (name == null) {
+            return null;
+        }
+        var config = Main.getMainConfig().getConfigurationSection("push-notification");
+        var section = config.getConfigurationSection(name);
+        if (section == null) {
+            return null;
+        }
+        switch (section.getString("type")) {
+            case "smtp" -> {
+                return new SmtpPushProvider(section);
+            }
+            case "pushplus" -> {
+                return new PushPlusPushProvider(section);
+            }
+            case "serverchan" -> {
+                return new ServerChanPushProvider(section);
+            }
+            case "telegram" -> {
+                return new TelegramPushProvider(section);
+            }
+        }
+        return null;
     }
 
     @Override
